Added JWT

+++ src/main/java/com/heraizen/ems/dao/EmployeeDaoReports.java
@@ -6,9 +6,9 @@ import com.heraizen.ems.domain.DEPT;
 import com.heraizen.ems.dto.EmployeeCountDTO;
 
 public interface EmployeeDaoReports {
-	public int getEmployeeCount(DEPT dept);
+	public long getEmployeeCount(DEPT dept);
 
-	public int totalEmployees();
+	public long totalEmployees();
 
 	public List<EmployeeCountDTO> employeeCountByDept();
 }
